Added support for configurable classes names, couple of bug fixes

+++ src/python_eval_client.js
@@ -6,7 +6,7 @@ const tmp = require('tmp');
 const exec = require('child_process').exec;
 
 const CNTK_CMD_TEMPLATE = "%s " + path.join(__dirname, 'frcnn_detector.py') + 
-                          ' --input %s --json-output %s --model %s'
+                          ' --input %s --json-output %s --model %s --cntk-path %s'
 
 function resolveCntkEnvDir(cntkInstallDir) {
     entries = fs.readdirSync(cntkInstallDir)
@@ -25,31 +25,56 @@ function getAndEnsureJsonTempDir() {
     return tmpDirPath;
 }
 
-function buildCntkCmd(directoryPath, cntkModelPath, cntkEnvDirPath, jsonFilePath) {
-    return util.format(CNTK_CMD_TEMPLATE, path.join(cntkEnvDirPath, 'python'), directoryPath, jsonFilePath, cntkModelPath)
+function buildCntkCmd(directoryPath, cntkModelPath, cntkEnvDirPath, jsonFilePath, cntkInstallDir) {
+    return util.format(CNTK_CMD_TEMPLATE, path.join(cntkEnvDirPath, 'python'), directoryPath, 
+                       jsonFilePath, cntkModelPath, cntkInstallDir)
 }
 
 function runCNTK(cntk_cmd, cb) {
     var proc = exec(cntk_cmd);
-    proc.on('exit', (err)=> {
-        cb(err);
+    
+    var error_data = '';
+    var output_data = '';
+
+    proc.stdout.on('data', function(data) {
+        output_data += data;
+    });
+
+    proc.stderr.on('data', function(data) {
+        error_data += data;
+    });
+
+    proc.on('exit', (exitCode)=> {
+        err = null;
+        if (exitCode != 0) {
+            errorMesage = util.format("CNTK Process failed with error code %d\nError output:%s",exitCode, error_data);
+            err = Error(errorMesage)
+        }
+        cb(err, output_data);
     });
 }
 
-function evalDirectoryImp(directoryPath, cntkModelPath, cntkEnvDirPath, jsonTempDir, cb) {
+function evalDirectoryImp(directoryPath, cntkModelPath, cntkEnvDirPath, jsonTempDir, cntkInstallDir, verbose, cb) {
     tmp.tmpName({dir : jsonTempDir, postfix: '.json'}, (err, jsonFilePath)=> {
         if (err) {
             return cb(err);
         }
         
-        var cmd = buildCntkCmd(directoryPath, cntkModelPath, cntkEnvDirPath, jsonFilePath);
-        runCNTK(cmd, (err) =>{
+        var cmd = buildCntkCmd(directoryPath, cntkModelPath, cntkEnvDirPath, jsonFilePath, cntkInstallDir);
+        runCNTK(cmd, (err, output_data) =>{
+            if (verbose) {
+                console.info('CNTK process output:')
+                console.info(output_data);
+            }
+
             if (err) {
                 return cb(err);
             }
             fs.readFile(jsonFilePath, (err, content) =>{
                 // call delete on the file and ignore the result since the json is in a temp directory anyway
-                fs.unlink(jsonFilePath);
+                if (fs.existsSync(jsonFilePath)) {
+                    fs.unlink(jsonFilePath);
+                }
 
                 if (err) {
                     return cb(err);
@@ -66,9 +91,10 @@ function evalDirectoryImp(directoryPath, cntkModelPath, cntkEnvDirPath, jsonTemp
     });
 }
 
-function EvalClient(cntkModelPath, cntkInstallDir) {
+function EvalClient(cntkModelPath, cntkInstallDir, verbose) {
     this.cntkInstallDir = cntkInstallDir;
     this.cntkModelPath = cntkModelPath;
+    this.verbose = !!verbose;
     this.cntkEnvDirPath = resolveCntkEnvDir(cntkInstallDir);
     // add to path..
     process.env.PATH = this.cntkEnvDirPath + ';' + process.env.PATH;
@@ -76,7 +102,8 @@ function EvalClient(cntkModelPath, cntkInstallDir) {
 
     this.evalDirectory = function(directoryPath, cb) {
         try {
-            evalDirectoryImp(directoryPath, this.cntkModelPath, this.cntkEnvDirPath, this.jsonTempDir, cb);
+            evalDirectoryImp(directoryPath, this.cntkModelPath, this.cntkEnvDirPath, 
+            this.jsonTempDir, this.cntkInstallDir, this.verbose, cb);
         }
         catch(e) {
             return cb(e);
